% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockDatasets.R
\name{availableMockDatasets}
\alias{availableMockDatasets}
\title{List the available datasets}
\usage{
availableMockDatasets()
}
\value{
A character vector with the available datasets.
}
\description{
List the available datasets
}
\examples{
library(omock)

availableMockDatasets()

}
