% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeCdmVersion.R
\name{changeCdmVersion}
\alias{changeCdmVersion}
\title{Function to change cdm version}
\usage{
changeCdmVersion(cdm, version = "5.4")
}
\arguments{
\item{cdm}{A \code{cdm_reference} object where you want to amend the cdm version}

\item{version}{cdm version to convert to e.g. version 5.4}
}
\value{
Returns the modified \code{cdm} object with updated version
}
\description{
Function to change cdm version
}
\examples{
library(omock)

# Create a mock CDM reference change to cdm version 5.3
cdm <- mockCdmReference()

cdm <- cdm |> changeCdmVersion(version = "5.4")

# View cdm version
print(cdmVersion(cdm))
}
