% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockDatasets.R
\name{downloadMockDataset}
\alias{downloadMockDataset}
\title{Download an OMOP Synthetic dataset.}
\usage{
downloadMockDataset(
  datasetName = "GiBleed",
  path = mockDatasetsFolder(),
  overwrite = NULL
)
}
\arguments{
\item{datasetName}{Name of the mock dataset. See `availableMockDatasets()`
for possibilities.}

\item{path}{Path where to download the dataset.}

\item{overwrite}{Whether to overwrite the dataset if it is already
downloaded. If NULL the used is asked whether to overwrite.}
}
\value{
The path to the downloaded dataset.
}
\description{
Download an OMOP Synthetic dataset.
}
\examples{
\donttest{
library(omock)

isMockDatasetDownloaded("GiBleed")
downloadMockDataset("GiBleed")
isMockDatasetDownloaded("GiBleed")
}

}
