% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockDatasets.R
\name{isMockDatasetDownloaded}
\alias{isMockDatasetDownloaded}
\title{Check if a certain dataset is downloaded.}
\usage{
isMockDatasetDownloaded(datasetName = "GiBleed", path = mockDatasetsFolder())
}
\arguments{
\item{datasetName}{Name of the mock dataset. See \code{availableMockDatasets()}
for possibilities.}

\item{path}{Path where to search for the dataset.}
}
\value{
Whether the dataset is available or not.
}
\description{
Check if a certain dataset is downloaded.
}
\examples{
\donttest{
library(omock)

isMockDatasetDownloaded("GiBleed")
downloadMockDataset("GiBleed")
isMockDatasetDownloaded("GiBleed")
}

}
