% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockCdmFromTables.R
\name{mockCdmFromTable}
\alias{mockCdmFromTable}
\title{mockCdmFrom Table
This function creates a mock cdm object from user self defined tables}
\usage{
mockCdmFromTable(cdm, cohortTable = NULL)
}
\arguments{
\item{cdm}{Name of the cdm object}

\item{cohortTable}{List of user self defined cohort table}
}
\value{
a cdm object
}
\description{
mockCdmFrom Table
This function creates a mock cdm object from user self defined tables
}
\examples{
library(omock)
}
