% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockCdmReference.R
\name{mockCdmReference}
\alias{mockCdmReference}
\title{Function to create empty cdm reference for mock database}
\usage{
mockCdmReference(cdmName = "mock database", vocabularySet = "mock")
}
\arguments{
\item{cdmName}{name of the mock cdm object}

\item{vocabularySet}{name of the vocabulary set}
}
\value{
an empty cdm object
}
\description{
Function to create empty cdm reference for mock database
}
\examples{
\donttest{
library(omock)
cdm <- mockCdmReference()

cdm

}
}
