% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockCdmReference.R
\name{mockCdmRefernce}
\alias{mockCdmRefernce}
\title{Function to create empty cdm reference for mock database}
\usage{
mockCdmRefernce(cdmName = "mock database")
}
\arguments{
\item{cdmName}{name of the mock cdm object}
}
\value{
an empty cdm object
}
\description{
Function to create empty cdm reference for mock database
}
\examples{
\donttest{
library(omock)
}
}
