% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockConcept.R
\name{mockConcept}
\alias{mockConcept}
\title{function to add mockconcept et to concept table}
\usage{
mockConcept(cdm, conceptSet, domain = "Condition", seed = 1)
}
\arguments{
\item{cdm}{Name of the cdm object}

\item{conceptSet}{conceptset as a vector}

\item{domain}{the domain of the conceptSet}

\item{seed}{random seed}
}
\value{
A cdm reference with the updated concept table
}
\description{
function to add mockconcept et to concept table
}
\examples{
library(omock)
library(dplyr)

cdm <- mockCdmReference() |> mockConcept(conceptSet = c(100,200), domain = "Condition")

cdm$concept |> filter(domain_id == "Condition")

}
