% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockDatasets.R
\docType{data}
\name{mockDatasets}
\alias{mockDatasets}
\title{Available mock OMOP CDM Synthetic Datasets}
\format{
A data frame with 4 variables:
\describe{
  \item{dataset_name}{Name of the dataset.}
  \item{url}{url to download the dataset.}
  \item{cdm_name}{Name of the cdm reference created.}
  \item{cdm_version}{OMOP CDM version of the dataset.}
  \item{size}{Size in bytes of the dataset.}
  \item{size_mb}{Size in Mega bytes of the dataset.}
}
}
\usage{
mockDatasets
}
\description{
These are the mock OMOP CDM Synthetic Datasets that are available to download
using the `omock` package.
}
\examples{
mockDatasets

}
\keyword{datasets}
