% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockObservationPeriod.R
\name{mockObservationPeriod}
\alias{mockObservationPeriod}
\title{Generates a mock observation period table and integrates it into an existing CDM object.}
\usage{
mockObservationPeriod(cdm, seed = NULL)
}
\arguments{
\item{cdm}{A local \code{cdm_reference} object used as the base structure to update.}

\item{seed}{An optional integer used to set the random seed for reproducibility. If \code{NULL}, the seed is not set.}
}
\value{
A modified \code{cdm_reference} object.
}
\description{
This function simulates observation periods for individuals based on their date of birth recorded in the 'person' table of the CDM object. It assigns random start and end dates for each observation period within a realistic timeframe up to a specified or default maximum date.
}
\examples{
library(omock)
library(dplyr)
# Create a mock CDM reference and add observation periods
cdm <- mockCdmReference() |>
  mockPerson(nPerson = 100) |>
  mockObservationPeriod()

# View the generated observation period data
cdm$observation_period |>
glimpse()
}
