% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockVisitOccurrence.R
\name{mockVisitOccurrence}
\alias{mockVisitOccurrence}
\title{Function to generate visit occurrence table}
\usage{
mockVisitOccurrence(cdm, seed = NULL, visitDetail = FALSE)
}
\arguments{
\item{cdm}{A local \code{cdm_reference} object used as the base structure to update.}

\item{seed}{An optional integer used to set the random seed for reproducibility. If \code{NULL}, the seed is not set.}

\item{visitDetail}{TRUE/FALSE it add the corresponding visit_detail table for
the mock visit occurrence created.}
}
\value{
A modified \code{cdm_reference} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
library(omock)

}
