% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockVocabulary.R
\name{mockVocabularyTable}
\alias{mockVocabularyTable}
\title{It creates a mock database with the vocabulary.}
\usage{
mockVocabularyTable(
  cdm,
  cdmSource = NULL,
  concept = NULL,
  vocabulary = NULL,
  conceptRelationship = NULL,
  conceptSynonym = NULL,
  conceptAncestor = NULL,
  drugStrength = NULL
)
}
\arguments{
\item{cdm}{name of the cdm object}

\item{cdmSource}{cdm source table.}

\item{concept}{Concept table.}

\item{vocabulary}{Vocabulary table}

\item{conceptRelationship}{Concept_relationship table.}

\item{conceptSynonym}{Concept_synonym table.}

\item{conceptAncestor}{Concept_ancestor table.}

\item{drugStrength}{Drug_strength table.}
}
\value{
A cdm reference with the vocabulary mock tables
}
\description{
It creates a mock database with the vocabulary.
}
\examples{
\donttest{
library(omock)
}

}
