% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omock-package.R
\docType{package}
\name{omock-package}
\alias{omock}
\alias{omock-package}
\title{omock: What the Package Does (Title Case)}
\description{
It creates a mock omop cdm object.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oxford-pharmacoepi.github.io/omock/}
}

}
\keyword{internal}
