% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockDatasets.R
\name{mockCdmFromDataset}
\alias{mockCdmFromDataset}
\title{Create a \code{local} cdm_reference from a dataset.}
\usage{
mockCdmFromDataset(
  datasetName = "GiBleed",
  source = "local",
  cdmVersion = NULL
)
}
\arguments{
\item{datasetName}{Name of the mock dataset. See \code{availableMockDatasets()} for supported values.}

\item{source}{Choice between \code{local} or \code{duckdb}.}

\item{cdmVersion}{Version of the OMOP CDM, can either be '5.3' or '5.4'. By
default, the dataset's original CDM version is used. If a different version
is requested, the returned CDM is converted with \code{changeCdmVersion()}.}
}
\value{
A local cdm_reference object.
}
\description{
Create a \code{local} cdm_reference from a dataset.
}
\examples{
library(omock)

mockDatasetsFolder(tempdir())
downloadMockDataset(datasetName = "GiBleed")
cdm <- mockCdmFromDataset(datasetName = "GiBleed")
cdm

}
