% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockDatasets.R
\name{mockCdmFromDataset}
\alias{mockCdmFromDataset}
\title{Create a `local` cdm_reference from a dataset.}
\usage{
mockCdmFromDataset(datasetName = "GiBleed")
}
\arguments{
\item{datasetName}{Name of the mock dataset. See `availableMockDatasets()`
for possibilities.}
}
\value{
A local cdm_reference object.
}
\description{
Create a `local` cdm_reference from a dataset.
}
\examples{
library(omock)

mockDatasetsFolder(tempdir())
downloadMockDataset(datasetName = "GiBleed")
cdm <- mockCdmFromDataset(datasetName = "GiBleed")
cdm

}
