% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockCdmFromTables.R
\name{mockCdmFromTables}
\alias{mockCdmFromTables}
\title{The goal of this function is to create a mock Cdm reference from user self defined tables.}
\usage{
mockCdmFromTables(cdm = mockCdmReference(), tables, seed = 1)
}
\arguments{
\item{cdm}{Name of the cdm object}

\item{seed}{random seed}

\item{cohortTable}{List of user self defined cohort table}
}
\value{
a cdm object
}
\description{
The goal of this function is to create a mock Cdm reference from user self defined tables.
}
\examples{
\donttest{
library(omock)
library(dplyr)

cohort <- tibble(
  cohort_definition_id = c(1, 1, 2, 2, 1, 3, 3, 3, 1, 3),
  subject_id = c(1, 4, 2, 3, 5, 5, 4, 3, 3, 1),
  cohort_start_date = as.Date(c(
    "2020-04-01", "2021-06-01", "2022-05-22", "2010-01-01", "2019-08-01",
    "2019-04-07", "2021-01-01", "2008-02-02", "2009-09-09", "2021-01-01"
  )),
  cohort_end_date = cohort_start_date
)

cdm <- mockCdmFromTables(tables = list(cohort = cohort))

cdm
cdm$cohort
cdm$person

}
}
