% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockCohort.R
\name{mockCohort}
\alias{mockCohort}
\title{Function to generate synthetic Cohort}
\usage{
mockCohort(
  cdm,
  tableName = "cohort",
  numberCohorts = 1,
  cohortName = paste0("cohort_", seq_len(numberCohorts)),
  recordPerson = 1,
  seed = 1
)
}
\arguments{
\item{cdm}{cdm object}

\item{tableName}{name of the cohort table}

\item{numberCohorts}{number of different cohort to create}

\item{cohortName}{name of the cohort within the table}

\item{recordPerson}{the expected number of record per person}

\item{seed}{random seed}
}
\value{
A cdm reference with the mock tables
}
\description{
Function to generate synthetic Cohort
}
\examples{
library(omock)
cdm <- mockCdmReference() |>
  mockPerson(nPerson = 100) |>
  mockObservationPeriod() |>
  mockCohort(
    tableName = "omock_example",
    numberCohorts = 2,
    cohortName = c("omock_cohort_1", "omock_cohort_2")
  )
}
