% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockConcept.R
\name{mockConcepts}
\alias{mockConcepts}
\title{Adds mock concept data to a concept table within a Common Data Model (CDM) object.}
\usage{
mockConcepts(cdm, conceptSet, domain = "Condition", seed = NULL)
}
\arguments{
\item{cdm}{A \code{cdm_reference} object used as the base structure to update.}

\item{conceptSet}{A numeric vector of concept IDs to be added or updated in
the concept table.These IDs should be unique within the
context of the provided domain to avoid unintended
overwriting unless that is the intended effect.}

\item{domain}{A character string specifying the domain of the concepts being
added.Only accepts "Condition", "Drug", "Measurement", or
"Observation". This defines under which category the concepts
fall and affects which vocabulary is used for them.}

\item{seed}{An optional integer used to set the random seed for reproducibility. If \code{NULL}, the seed is not set.}
}
\value{
A modified \code{cdm_reference} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This function inserts new concept entries into a specified domain within
the concept table of a CDM object.It supports four domains: Condition, Drug,
Measurement, and Observation. Existing entries with the same concept IDs
will be overwritten, so caution should be used when adding data to prevent
unintended data loss.
}
\examples{
library(omock)
library(dplyr)

# Create a mock CDM reference and add concepts in the 'Condition' domain
cdm <- mockCdmReference() |> mockConcepts(
  conceptSet = c(100, 200), domain = "Condition"
)

# View the updated concept entries for the 'Condition' domain
cdm$concept |> filter(domain_id == "Condition")

}
