% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockObservationPeriod.R
\name{mockObservationPeriod}
\alias{mockObservationPeriod}
\title{mockObservationPeriod}
\usage{
mockObservationPeriod(cdm, seed = 1)
}
\arguments{
\item{cdm}{Name of the cdm object}

\item{seed}{random seed}
}
\value{
a cdm object with mock observation table
}
\description{
mockObservationPeriod
}
\examples{
\donttest{
library(omock)
cdm <- mockCdmReference() |>
  mockPerson(nPerson = 100) |>
  mockObservationPeriod()
}
}
