% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockVisitOccurrence.R
\name{mockVisitOccurrence}
\alias{mockVisitOccurrence}
\title{Function to generate visit occurrence table}
\usage{
mockVisitOccurrence(cdm, seed = NULL, visitDetail = FALSE)
}
\arguments{
\item{cdm}{the CDM reference into which the  mock visit occurrence table will
be added}

\item{seed}{A random seed to ensure reproducibility of the generated data.}

\item{visitDetail}{TRUE/FALSE it add the corresponding visit_detail table for
the mock visit occurrence created.}
}
\value{
A cdm reference with the visit_occurrence tables added
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
library(omock)

}
