% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockVisitOccurrence.R
\name{mockVisitOccurrence}
\alias{mockVisitOccurrence}
\title{Function to generate visit occurrence table}
\usage{
mockVisitOccurrence(cdm, seed = NULL)
}
\arguments{
\item{cdm}{the CDM reference into which the  mock visit occurrence table will be added}

\item{seed}{A random seed to ensure reproducibility of the generated data.}
}
\value{
A cdm reference with the visit_occurrence tables added
}
\description{
Function to generate visit occurrence table
}
\examples{
library(omock)

}
