% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omock-package.R
\docType{package}
\name{omock-package}
\alias{omock}
\alias{omock-package}
\title{omock: Creation of Mock Observational Medical Outcomes Partnership Common Data Model}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Creates mock data for testing and package development for the Observational Medical Outcomes Partnership common data model. The package offers functions crafted with pipeline-friendly implementation, enabling users to effortlessly include only the necessary tables for their testing needs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oxford-pharmacoepi.github.io/omock/}
  \item \url{https://github.com/oxford-pharmacoepi/omock}
  \item Report bugs at \url{https://github.com/oxford-pharmacoepi/omock/issues}
}

}
\author{
\strong{Maintainer}: Mike Du \email{mike.du@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9517-8834}{ORCID})

Authors:
\itemize{
  \item Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})
  \item Edward Burn \email{edward.burn@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9286-1128}{ORCID})
  \item Nuria Mercade-Besora \email{nuria.mercadebesora@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0006-7948-3747}{ORCID})
  \item Xihang Chen \email{xihang.chen@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0001-8112-8959}{ORCID})
}

}
\keyword{internal}
