% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-omopgenerics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{emptyCdmReference}
\alias{cdmName}
\alias{cdmVersion}
\alias{cdmSource}
\alias{dropSourceTable}
\alias{readSourceTable}
\alias{listSourceTables}
\alias{insertTable}
\alias{omopDataFolder}
\alias{cdmReference}
\alias{tableName}
\alias{tableSource}
\alias{omopColumns}
\alias{omopTables}
\alias{cohortColumns}
\alias{cohortTables}
\alias{achillesColumns}
\alias{achillesTables}
\alias{cdmDisconnect}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{omopgenerics}{\code{\link[omopgenerics]{achillesColumns}}, \code{\link[omopgenerics]{achillesTables}}, \code{\link[omopgenerics]{cdmDisconnect}}, \code{\link[omopgenerics]{cdmName}}, \code{\link[omopgenerics]{cdmReference}}, \code{\link[omopgenerics]{cdmSource}}, \code{\link[omopgenerics]{cdmVersion}}, \code{\link[omopgenerics]{cohortColumns}}, \code{\link[omopgenerics]{cohortTables}}, \code{\link[omopgenerics]{dropSourceTable}}, \code{\link[omopgenerics]{dropSourceTable}}, \code{\link[omopgenerics]{emptyCdmReference}}, \code{\link[omopgenerics]{insertTable}}, \code{\link[omopgenerics]{listSourceTables}}, \code{\link[omopgenerics]{listSourceTables}}, \code{\link[omopgenerics]{omopColumns}}, \code{\link[omopgenerics]{omopDataFolder}}, \code{\link[omopgenerics]{omopTables}}, \code{\link[omopgenerics]{readSourceTable}}, \code{\link[omopgenerics]{readSourceTable}}, \code{\link[omopgenerics]{tableName}}, \code{\link[omopgenerics]{tableSource}}}
}}

